+++ footer.php
@@ -2,7 +2,7 @@
 /**
  * The template for displaying the footer.
  * @since WP-Forge 5.5.1.7
- * @version 6.3.0.1
+ * @version 6.3.0.2
  */
 ?>
 	</section><!-- end .content-wrap -->