+++ page-templates/full-width.php
@@ -3,7 +3,7 @@
  * Template Name: Full-width Page Template, No Sidebar
  * Template Post Type: post, page, aside, audio, chat, gallery, image, link, quote, status, video
  * @since WP-Forge 5.5.1.7
- * @version 6.3.0.1
+ * @version 6.3.0.2
  */
 get_header(); ?>
 	<div id="content" class="small-12 large-12 columns" role="main">