+++ searchform.php
@@ -2,7 +2,7 @@
 /**
  * The template for displaying the search form.
  * @since WP-Forge 5.5.1.7
- * @version 6.3.0.1
+ * @version 6.3.0.2
  */
 ?>
 <form role="search" method="get" id="searchform" action="<?php echo home_url('/'); ?>">